+++ src/lib/components/document/edit-page/edit-items/EditMenu.tsx
@@ -107,8 +107,6 @@ const EditMenu = () => {
 
   const handleOpenGroups = () => {
     if (editStage === EDIT_STAGES.DONE) {
-      setEditStage(EDIT_STAGES.NULL)
-      setGroupId('')
       setGroupsModalVisibility(true)
     }
   }